% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert.R
\name{Alert}
\alias{Alert}
\title{Odbc}
\description{
Creates an Odbc class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Alert-send}{\code{Alert$send()}}
\item \href{#method-Alert-clone}{\code{Alert$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Alert-send"></a>}}
\if{latex}{\out{\hypertarget{method-Alert-send}{}}}
\subsection{Method \code{send()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Alert$send(
  text,
  url = Sys.getenv("SLACK_URL"),
  token = Sys.getenv("SLACK_TOKEN"),
  channel = Sys.getenv("SLACK_CHANNEL"),
  username = Sys.getenv("SLACK_USERNAME"),
  emoji = Sys.getenv("SLACK_EMOJI", ":bird:")
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Alert-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Alert-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Alert$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
