% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{Connection}
\alias{Connection}
\title{Connection}
\description{
Creates an Connection class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Connection-new}{\code{Connection$new()}}
\item \href{#method-Connection-open}{\code{Connection$open()}}
\item \href{#method-Connection-close}{\code{Connection$close()}}
\item \href{#method-Connection-consult}{\code{Connection$consult()}}
\item \href{#method-Connection-clone}{\code{Connection$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Connection-new"></a>}}
\if{latex}{\out{\hypertarget{method-Connection-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Connection$new(driver, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Connection-open"></a>}}
\if{latex}{\out{\hypertarget{method-Connection-open}{}}}
\subsection{Method \code{open()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Connection$open(
  driver,
  ...,
  min_size = 1L,
  max_size = 5L,
  on_create = NULL,
  idle_timeout = 5L * 60L,
  validation_interval = 1L * 60L,
  validate_query = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Connection-close"></a>}}
\if{latex}{\out{\hypertarget{method-Connection-close}{}}}
\subsection{Method \code{close()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Connection$close()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Connection-consult"></a>}}
\if{latex}{\out{\hypertarget{method-Connection-consult}{}}}
\subsection{Method \code{consult()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Connection$consult(query)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Connection-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Connection-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Connection$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
