% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{Error}
\alias{Error}
\title{Error}
\description{
Creates a Error class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Error-check}{\code{Error$check()}}
\item \href{#method-Error-clone}{\code{Error$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Error-check"></a>}}
\if{latex}{\out{\hypertarget{method-Error-check}{}}}
\subsection{Method \code{check()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Error$check()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Error-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Error-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Error$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
