% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthr-package.R
\docType{package}
\name{healthr-package}
\alias{healthr}
\alias{healthr-package}
\title{healthr: Functions for Health Tracing}
\description{
Functions for Health Tracing in OCC standarizing classes and methods
}
\author{
\strong{Maintainer}: Abraham Jaimes \email{ajaimes@occ.com.mx}

}
\keyword{internal}
