% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odbc.R
\name{Odbc}
\alias{Odbc}
\title{Odbc}
\description{
Creates an Odbc class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Odbc-new}{\code{Odbc$new()}}
\item \href{#method-Odbc-connect}{\code{Odbc$connect()}}
\item \href{#method-Odbc-disconnect}{\code{Odbc$disconnect()}}
\item \href{#method-Odbc-consult}{\code{Odbc$consult()}}
\item \href{#method-Odbc-clone}{\code{Odbc$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Odbc-new"></a>}}
\if{latex}{\out{\hypertarget{method-Odbc-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Odbc$new(
  driver = "ODBC Driver 17 for SQL Server",
  host = Sys.getenv("DB_HOST"),
  name = Sys.getenv("DB_NAME"),
  user = Sys.getenv("DB_USER"),
  password = Sys.getenv("DB_PASSWORD"),
  extra = "ApplicationIntent=ReadOnly"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Odbc-connect"></a>}}
\if{latex}{\out{\hypertarget{method-Odbc-connect}{}}}
\subsection{Method \code{connect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Odbc$connect(
  driver = "ODBC Driver 17 for SQL Server",
  host = Sys.getenv("DB_HOST"),
  name = Sys.getenv("DB_NAME"),
  user = Sys.getenv("DB_USER"),
  password = Sys.getenv("DB_PASSWORD"),
  extra = "ApplicationIntent=ReadOnly"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Odbc-disconnect"></a>}}
\if{latex}{\out{\hypertarget{method-Odbc-disconnect}{}}}
\subsection{Method \code{disconnect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Odbc$disconnect()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Odbc-consult"></a>}}
\if{latex}{\out{\hypertarget{method-Odbc-consult}{}}}
\subsection{Method \code{consult()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Odbc$consult(query)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Odbc-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Odbc-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Odbc$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
