% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{Time}
\alias{Time}
\title{Time}
\description{
Creates a Time class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Time-new}{\code{Time$new()}}
\item \href{#method-Time-scale}{\code{Time$scale()}}
\item \href{#method-Time-clone}{\code{Time$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Time-new"></a>}}
\if{latex}{\out{\hypertarget{method-Time-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Time$new(add = 6L * 60L * 60L)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Time-scale"></a>}}
\if{latex}{\out{\hypertarget{method-Time-scale}{}}}
\subsection{Method \code{scale()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Time$scale(by = 1L, horizont = 1440L)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Time-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Time-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Time$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
