% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{Metric}
\alias{Metric}
\title{Metric}
\description{
Creates a Metric class
}
\section{Super class}{
\code{\link[healthr:Parameter]{healthr::Parameter}} -> \code{Metric}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Metric-new}{\code{Metric$new()}}
\item \href{#method-Metric-count}{\code{Metric$count()}}
\item \href{#method-Metric-clone}{\code{Metric$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="healthr" data-topic="Parameter" data-id="set"><a href='../../healthr/html/Parameter.html#method-Parameter-set'><code>healthr::Parameter$set()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metric-new"></a>}}
\if{latex}{\out{\hypertarget{method-Metric-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metric$new(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metric-count"></a>}}
\if{latex}{\out{\hypertarget{method-Metric-count}{}}}
\subsection{Method \code{count()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metric$count(dt, param)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metric-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Metric-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metric$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
