% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{Model}
\alias{Model}
\title{Model}
\description{
Creates a Model class
}
\section{Super classes}{
\code{\link[healthr:Parameter]{healthr::Parameter}} -> \code{\link[healthr:Metric]{healthr::Metric}} -> \code{Model}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Model-new}{\code{Model$new()}}
\item \href{#method-Model-normalize}{\code{Model$normalize()}}
\item \href{#method-Model-train}{\code{Model$train()}}
\item \href{#method-Model-predict}{\code{Model$predict()}}
\item \href{#method-Model-clone}{\code{Model$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="healthr" data-topic="Parameter" data-id="set"><a href='../../healthr/html/Parameter.html#method-Parameter-set'><code>healthr::Parameter$set()</code></a></span></li>
<li><span class="pkg-link" data-pkg="healthr" data-topic="Metric" data-id="count"><a href='../../healthr/html/Metric.html#method-Metric-count'><code>healthr::Metric$count()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Model-new"></a>}}
\if{latex}{\out{\hypertarget{method-Model-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$new(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Model-normalize"></a>}}
\if{latex}{\out{\hypertarget{method-Model-normalize}{}}}
\subsection{Method \code{normalize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$normalize(dt, param)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Model-train"></a>}}
\if{latex}{\out{\hypertarget{method-Model-train}{}}}
\subsection{Method \code{train()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$train(k = 4L)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Model-predict"></a>}}
\if{latex}{\out{\hypertarget{method-Model-predict}{}}}
\subsection{Method \code{predict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$predict(param, level = 99L, k = 4L)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Model-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Model-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
