% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redis.R
\name{Redis}
\alias{Redis}
\title{Redis}
\description{
Creates a Redis class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Redis-new}{\code{Redis$new()}}
\item \href{#method-Redis-connect}{\code{Redis$connect()}}
\item \href{#method-Redis-get}{\code{Redis$get()}}
\item \href{#method-Redis-set}{\code{Redis$set()}}
\item \href{#method-Redis-clone}{\code{Redis$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Redis-new"></a>}}
\if{latex}{\out{\hypertarget{method-Redis-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Redis$new(
  host = Sys.getenv("REDIS_HOST"),
  port = Sys.getenv("REDIS_PORT"),
  db = Sys.getenv("REDIS_DB")
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Redis-connect"></a>}}
\if{latex}{\out{\hypertarget{method-Redis-connect}{}}}
\subsection{Method \code{connect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Redis$connect(host, port, db)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Redis-get"></a>}}
\if{latex}{\out{\hypertarget{method-Redis-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Redis$get(key)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Redis-set"></a>}}
\if{latex}{\out{\hypertarget{method-Redis-set}{}}}
\subsection{Method \code{set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Redis$set(key, value, EX = NULL, PX = NULL, condition = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Redis-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Redis-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Redis$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
